test: merge suspend and resume authorization tests
ProcessInstanceSuspendAuthorizationTest and
ProcessInstanceResumeAuthorizationTest were near-duplicates gating on the
same SUSPEND_PROCESS_INSTANCE permission, with identical setup and helper
methods. Merged into one class to remove the duplication.
